@@ -2,29 +2,34 @@
   "name": "wand-investments",
   "version": "1.0.0",
   "private": true,
+  "author": "Garen Crowngaurd",
   "homepage": "https://GarenCrownguard.github.io/",
+  "description": "Dapp for Wand Investments",
   "dependencies": {
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-burger-menu": "^2.6.13",
-    "react-collapsible-content": "^1.0.0",
-    "react-dom": "^16.8.6",
-    "react-jss": "^10.4.0",
+    "@chakra-ui/icons": "^2.0.4",
+    "@chakra-ui/react": "^2.2.4",
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "axios": "^0.27.2",
+    "babel-loader": "8.1.0",
+    "concurrently": "^7.3.0",
+    "css-loader": "^6.7.1",
+    "framer-motion": "^4.1.17",
+    "json-loader": "^0.5.7",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-redux": "^8.0.2",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "react-svg-line-chart": "^2.0.1",
+    "react-scripts": "4.0.3",
     "redux": "^4.2.0",
-    "redux-devtools-extension": "^2.13.9",
-    "simple-flexbox": "^2.3.1"
+    "redux-devtools-extension": "^2.13.9"
   },
   "scripts": {
     "start": "react-scripts start",
+    "start-api": "node api/index.js",
     "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d build"
+    "deploy": "gh-pages -d build",
+    "dev": "concurrently --kill-others \"npm start\" \"npm run start-api\""
   },
   "eslintConfig": {
     "extends": [
@@ -37,11 +42,7 @@
       "not dead",
       "not op_mini all"
     ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "development": []
   },
   "devDependencies": {
     "gh-pages": "^4.0.0"
