@@ -1,35 +1,32 @@
-import IconBurger from "./icon-burger";
+// import IconBurger from "./icon-burger";
 
-import IconLogo from "./icon-logo";
-import IconSoon from "./icon-soon";
-import IconWallet from "./icon-wallet";
-import IconSwap from "./icon-swap";
-import IconVerticalseperator from "./icon-verticalseperator";
+// import IconLogo from "./icon-logo";
+// import IconSoon from "./icon-soon";
+// import IconWallet from "./icon-wallet";
+// import IconSwap from "./icon-swap";
+// import IconVerticalseperator from "./icon-verticalseperator";
 
-import iconAccount from "./Icon-account";
-import iconBridge from "./Icon-bridge";
-import iconDashboard from "./Icon-dashboard";
-import iconDocs from "./Icon-docs";
-import iconGovernance from "./Icon-governance";
-import iconLending from "./Icon-lending";
-import iconSwapSidebar from "./Icon-swap-sidebar";
-import iconVote from "./Icon-vote";
+// import iconBridge from "./icon-bridge";
+// import iconDashboard from "./icon-dashboard";
+// import iconDocs from "./icon-docs";
+// import iconGovernance from "./icon-governance";
+// import iconLending from "./icon-lending";
+// import iconSwapSidebar from "./icon-swap-sidebar";
+// 
+// export {
+//   IconBurger,
 
-export {
-  IconBurger,
   
-  IconLogo,
-  IconSoon,
-  IconWallet,
-  IconSwap,
-  IconVerticalseperator,
+//   IconLogo,
+//   IconSoon,
+//   IconWallet,
+//   IconSwap,
+//   IconVerticalseperator,
 
-  iconAccount,
-  iconBridge,
-  iconDashboard,
-  iconDocs,
-  iconGovernance,
-  iconLending,
-  iconSwapSidebar,
-  iconVote
-};
+//   iconBridge,
+//   iconDashboard,
+//   iconDocs,
+//   iconGovernance,
+//   iconLending,
+//   iconSwapSidebar,
+// };
