@@ -1,17 +1,5 @@
-/* Typography */
-
-export const fontFamily = 11;
-export const fontSize = 24;
-
-
-export const ACTION_INVALID = "This action is invalid!"
-export const CONSTANT_NUMBER_1 = 'hello I am a constant';
-export const CONSTANT_NUMBER_2 = 'hello I am also a constant';
+export const API_URL = "http://localhost:8448"
 
 /*
-
-import * as myConstClass from 'path/to/fileWithConstants';
-
-const errorMsg = myConstClass.ACTION_INVALID;
-
+import * as myConstants from 'path/to/fileWithConstants';
 */
